+++ R/wic_col4.R
@@ -1,4 +1,4 @@
-#' Colours used in Wittgenstein Human Capital Data Explorer
+#' Colours used in Wittgenstein Centre for Demography and Human Capital Data Explorer
 #'
 #' Three sets of colours used for filling education based plots based on the different availability of detailed education categories (four, six or eight groups)
 #'