+++ NEWS.md
@@ -14,3 +14,4 @@
 * Add start-up citation note
 * Renamed `wcde_pull()` to `get_wcde_single()`
 * Initial pkgdown working
+* Github actions working