+++ NEWS.md
@@ -15,3 +15,4 @@
 * Renamed `wcde_pull()` to `get_wcde_single()`
 * Initial pkgdown working
 * Github actions working
+* Website to description